+++ CHANGELOG.md
@@ -4,6 +4,26 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.7.2] - 2021-01-21
+### Fixed
+- Correctly set `dag.description` depending on Airflow version
+
+## [0.7.1] - 2020-12-19
+### Added
+- Examples for using Custom Operator
+### Fixed
+- Handle `"None"` as `schedule_interval`
+
+## [0.7.0] - 2020-12-19
+### Added
+- Support Airflow 2.0!
+
+## [0.6.0] - 2020-11-16
+### Added
+- `catchup` added to DAG parameters
+- Support for `ExternalTaskSensor`
+- Run test suite against Python 3.8
+
 ## [0.5.0] - 2020-08-20
 ### Added
 - Support for `KubernetesPodOperator`
@@ -70,7 +90,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.1.0] - 2018-11-20
 - Initial release
 
-[Unreleased]: https://github.com/ajbosco/dag-factory/compare/v0.5.0...HEAD
+[Unreleased]: https://github.com/ajbosco/dag-factory/compare/v0.7.2...HEAD
+[0.7.2]: https://github.com/ajbosco/dag-factory/compare/v0.7.1...v0.7.2
+[0.7.1]: https://github.com/ajbosco/dag-factory/compare/v0.7.0...v0.7.1
+[0.7.0]: https://github.com/ajbosco/dag-factory/compare/v0.6.0...v0.7.0
+[0.6.0]: https://github.com/ajbosco/dag-factory/compare/v0.5.0...v0.6.0
 [0.5.0]: https://github.com/ajbosco/dag-factory/compare/v0.4.5...v0.5.0
 [0.4.5]: https://github.com/ajbosco/dag-factory/compare/v0.4.4...v0.4.5
 [0.4.4]: https://github.com/ajbosco/dag-factory/compare/v0.4.3...v0.4.4